requirements.txt
boto3==1.20.34
botocore==1.23.34
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
mysql-connector==2.2.9
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
